requirements.txt
torch
torchvision
torchaudio
transformers
einops
opencv-python
pillow
pyside6
qdarktheme
deep-translator
